+++ R/unbiased-package.R
@@ -1,0 +1,9 @@
+#' @import checkmate
+#' @import dplyr
+#'
+#' @keywords internal
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL