+++ NEWS.md
@@ -1,3 +1,5 @@
 # unbiased (development version)
 
 * Initialized package structure.
+* Implemented study definition endpoint (`POST /study`).
+* Implemented study details endpoints (`GET /study`, `GET /study/<study_id>`).